Moved header content out of shared styles and wrap research header
The header before was a column on the phone in the research tab, so instead I've made it wrap. Might need a little bit more fine tuning but should be good for now. Also moved the header out of the shared styles and back to main so that both could could be controlled individually.